+++ client/src/featureform/resources.py
@@ -6,16 +6,15 @@ import sys
 import json
 import time
 import base64
-from enum import Enum
-from typeguard import typechecked
 from typing import List, Tuple, Union, Optional
 
 import dill
 import grpc
+
+from .resources_config import Config, EmptyConfig
 from .sqlite_metadata import SQLiteMetadata
 from google.protobuf.duration_pb2 import Duration
 
-from featureform.proto import metadata_pb2 as pb
 from dataclasses import dataclass, field
 from .version import check_up_to_date
 from .exceptions import *
@@ -62,519 +61,6 @@ class Schedule:
 
 @typechecked
 @dataclass
-class RedisConfig:
-    host: str
-    port: int
-    password: str
-    db: int
-
-    def software(self) -> str:
-        return "redis"
-
-    def type(self) -> str:
-        return "REDIS_ONLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Addr": f"{self.host}:{self.port}",
-            "Password": self.password,
-            "DB": self.db,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class AWSCredentials:
-    def __init__(
-        self,
-        aws_access_key_id: str = "",
-        aws_secret_access_key: str = "",
-    ):
-        empty_strings = aws_access_key_id == "" or aws_secret_access_key == ""
-        if empty_strings:
-            raise Exception(
-                "'AWSCredentials' requires all parameters: 'aws_access_key_id', 'aws_secret_access_key'"
-            )
-
-        self.aws_access_key_id = aws_access_key_id
-        self.aws_secret_access_key = aws_secret_access_key
-
-    def type(self):
-        return "AWS_CREDENTIALS"
-
-    def config(self):
-        return {
-            "AWSAccessKeyId": self.aws_access_key_id,
-            "AWSSecretKey": self.aws_secret_access_key,
-        }
-
-
-@typechecked
-@dataclass
-class GCPCredentials:
-    def __init__(
-        self,
-        project_id: str,
-        credentials_path: str,
-    ):
-        self.project_id = project_id
-        self.credentials = json.load(open(credentials_path))
-
-    def type(self):
-        return "GCPCredentials"
-
-    def config(self):
-        return {
-            "ProjectId": self.project_id,
-            "JSON": self.credentials,
-        }
-
-
-@typechecked
-@dataclass
-class GCSFileStoreConfig:
-    credentials: GCPCredentials
-    bucket_name: str
-    bucket_path: str = ""
-
-    def software(self) -> str:
-        return "gcs"
-
-    def type(self) -> str:
-        return "GCS"
-
-    def serialize(self) -> bytes:
-        config = {
-            "BucketName": self.bucket_name,
-            "BucketPath": self.bucket_path,
-            "Credentials": self.credentials.config(),
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-    def config(self):
-        return {
-            "BucketName": self.bucket_name,
-            "BucketPath": self.bucket_path,
-            "Credentials": self.credentials.config(),
-        }
-
-    def store_type(self):
-        return self.type()
-
-
-@typechecked
-@dataclass
-class AzureFileStoreConfig:
-    account_name: str
-    account_key: str
-    container_name: str
-    root_path: str
-
-    def software(self) -> str:
-        return "azure"
-
-    def type(self) -> str:
-        return "AZURE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "AccountName": self.account_name,
-            "AccountKey": self.account_key,
-            "ContainerName": self.container_name,
-            "Path": self.root_path,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-    def config(self):
-        return {
-            "AccountName": self.account_name,
-            "AccountKey": self.account_key,
-            "ContainerName": self.container_name,
-            "Path": self.root_path,
-        }
-
-    def store_type(self):
-        return self.type()
-
-
-@typechecked
-@dataclass
-class S3StoreConfig:
-    def __init__(
-        self,
-        bucket_path: str,
-        bucket_region: str,
-        credentials: AWSCredentials,
-        path: str = "",
-    ):
-        bucket_path_ends_with_slash = len(bucket_path) != 0 and bucket_path[-1] == "/"
-
-        if bucket_path_ends_with_slash:
-            raise Exception("The 'bucket_path' cannot end with '/'.")
-
-        self.bucket_path = bucket_path
-        self.bucket_region = bucket_region
-        self.credentials = credentials
-        self.path = path
-
-    def software(self) -> str:
-        return "S3"
-
-    def type(self) -> str:
-        return "S3"
-
-    def serialize(self) -> bytes:
-        config = self.config()
-        return bytes(json.dumps(config), "utf-8")
-
-    def config(self):
-        return {
-            "Credentials": self.credentials.config(),
-            "BucketRegion": self.bucket_region,
-            "BucketPath": self.bucket_path,
-            "Path": self.path,
-        }
-
-    def store_type(self):
-        return self.type()
-
-
-@typechecked
-@dataclass
-class HDFSConfig:
-    def __init__(self, host: str, port: str, path: str, username: str):
-        bucket_path_ends_with_slash = len(path) != 0 and path[-1] == "/"
-
-        if bucket_path_ends_with_slash:
-            raise Exception("The 'bucket_path' cannot end with '/'.")
-
-        self.path = path
-        self.host = host
-        self.port = port
-        self.username = username
-
-    def software(self) -> str:
-        return "HDFS"
-
-    def type(self) -> str:
-        return "HDFS"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Host": self.host,
-            "Port": self.port,
-            "Path": self.path,
-            "Username": self.username,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-    def config(self):
-        return {
-            "Host": self.host,
-            "Port": self.port,
-            "Path": self.path,
-            "Username": self.username,
-        }
-
-    def store_type(self):
-        return self.type()
-
-
-@typechecked
-@dataclass
-class OnlineBlobConfig:
-    store_type: str
-    store_config: dict
-
-    def software(self) -> str:
-        return self.store_type
-
-    def type(self) -> str:
-        return "BLOB_ONLINE"
-
-    def config(self):
-        return self.store_config
-
-    def serialize(self) -> bytes:
-        config = {
-            "Type": self.store_type,
-            "Config": self.store_config,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class FirestoreConfig:
-    collection: str
-    project_id: str
-    credentials_path: str
-
-    def software(self) -> str:
-        return "firestore"
-
-    def type(self) -> str:
-        return "FIRESTORE_ONLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Collection": self.collection,
-            "ProjectID": self.project_id,
-            "Credentials": json.load(open(self.credentials_path)),
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class CassandraConfig:
-    keyspace: str
-    host: str
-    port: str
-    username: str
-    password: str
-    consistency: str
-    replication: int
-
-    def software(self) -> str:
-        return "cassandra"
-
-    def type(self) -> str:
-        return "CASSANDRA_ONLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Keyspace": self.keyspace,
-            "Addr": f"{self.host}:{self.port}",
-            "Username": self.username,
-            "Password": self.password,
-            "Consistency": self.consistency,
-            "Replication": self.replication,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class DynamodbConfig:
-    region: str
-    access_key: str
-    secret_key: str
-
-    def software(self) -> str:
-        return "dynamodb"
-
-    def type(self) -> str:
-        return "DYNAMODB_ONLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Region": self.region,
-            "AccessKey": self.access_key,
-            "SecretKey": self.secret_key,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class MongoDBConfig:
-    username: str
-    password: str
-    host: str
-    port: str
-    database: str
-    throughput: int
-
-    def software(self) -> str:
-        return "mongodb"
-
-    def type(self) -> str:
-        return "MONGODB_ONLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Username": self.username,
-            "Password": self.password,
-            "Host": self.host,
-            "Port": self.port,
-            "Database": self.database,
-            "Throughput": self.throughput,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class LocalConfig:
-    def software(self) -> str:
-        return "localmode"
-
-    def type(self) -> str:
-        return "LOCAL_ONLINE"
-
-    def serialize(self) -> bytes:
-        config = {}
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class SnowflakeConfig:
-    username: str
-    password: str
-    schema: str
-    account: str = ""
-    organization: str = ""
-    database: str = ""
-    account_locator: str = ""
-    warehouse: str = ""
-    role: str = ""
-
-    def __post_init__(self):
-        if self.__has_legacy_credentials() and self.__has_current_credentials():
-            raise ValueError(
-                "Cannot create configure Snowflake with both current and legacy credentials"
-            )
-
-        if not self.__has_legacy_credentials() and not self.__has_current_credentials():
-            raise ValueError("Cannot create configure Snowflake without credentials")
-
-    def __has_legacy_credentials(self) -> bool:
-        return self.account_locator != ""
-
-    def __has_current_credentials(self) -> bool:
-        if (self.account != "" and self.organization == "") or (
-            self.account == "" and self.organization != ""
-        ):
-            raise ValueError("Both Snowflake organization and account must be included")
-        elif self.account != "" and self.organization != "":
-            return True
-        else:
-            return False
-
-    def software(self) -> str:
-        return "Snowflake"
-
-    def type(self) -> str:
-        return "SNOWFLAKE_OFFLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Username": self.username,
-            "Password": self.password,
-            "Organization": self.organization,
-            "AccountLocator": self.account_locator,
-            "Account": self.account,
-            "Schema": self.schema,
-            "Database": self.database,
-            "Warehouse": self.warehouse,
-            "Role": self.role,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class PostgresConfig:
-    host: str
-    port: str
-    database: str
-    user: str
-    password: str
-
-    def software(self) -> str:
-        return "postgres"
-
-    def type(self) -> str:
-        return "POSTGRES_OFFLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Host": self.host,
-            "Port": self.port,
-            "Username": self.user,
-            "Password": self.password,
-            "Database": self.database,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class RedshiftConfig:
-    host: str
-    port: str
-    database: str
-    user: str
-    password: str
-
-    def software(self) -> str:
-        return "redshift"
-
-    def type(self) -> str:
-        return "REDSHIFT_OFFLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "Host": self.host,
-            "Port": self.port,
-            "Username": self.user,
-            "Password": self.password,
-            "Database": self.database,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class BigQueryConfig:
-    project_id: str
-    dataset_id: str
-    credentials_path: str
-
-    def software(self) -> str:
-        return "bigquery"
-
-    def type(self) -> str:
-        return "BIGQUERY_OFFLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "ProjectID": self.project_id,
-            "DatasetID": self.dataset_id,
-            "Credentials": json.load(open(self.credentials_path)),
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class SparkConfig:
-    executor_type: str
-    executor_config: dict
-    store_type: str
-    store_config: dict
-
-    def software(self) -> str:
-        return "spark"
-
-    def type(self) -> str:
-        return "SPARK_OFFLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "ExecutorType": self.executor_type,
-            "StoreType": self.store_type,
-            "ExecutorConfig": self.executor_config,
-            "StoreConfig": self.store_config,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
 class K8sResourceSpecs:
     cpu_request: str = ""
     cpu_limit: str = ""
@@ -598,62 +84,6 @@ class K8sArgs:
             )
             transformation.kubernetes_args.specs.memory_limit = self.specs.memory_limit
         return transformation
-
-
-@typechecked
-@dataclass
-class K8sConfig:
-    store_type: str
-    store_config: dict
-    docker_image: str = ""
-
-    def software(self) -> str:
-        return "k8s"
-
-    def type(self) -> str:
-        return "K8S_OFFLINE"
-
-    def serialize(self) -> bytes:
-        config = {
-            "ExecutorType": "K8S",
-            "ExecutorConfig": {"docker_image": self.docker_image},
-            "StoreType": self.store_type,
-            "StoreConfig": self.store_config,
-        }
-        return bytes(json.dumps(config), "utf-8")
-
-
-@typechecked
-@dataclass
-class EmptyConfig:
-    def software(self) -> str:
-        return ""
-
-    def type(self) -> str:
-        return ""
-
-    def serialize(self) -> bytes:
-        return bytes("", "utf-8")
-
-
-Config = Union[
-    RedisConfig,
-    SnowflakeConfig,
-    PostgresConfig,
-    RedshiftConfig,
-    LocalConfig,
-    BigQueryConfig,
-    FirestoreConfig,
-    SparkConfig,
-    OnlineBlobConfig,
-    AzureFileStoreConfig,
-    S3StoreConfig,
-    K8sConfig,
-    MongoDBConfig,
-    GCSFileStoreConfig,
-    EmptyConfig,
-    HDFSConfig,
-]
 
 
 @typechecked
