+++ client/src/featureform/sqlite_metadata.py
@@ -253,8 +253,7 @@ class SQLiteMetadata:
           name             text PRIMARY KEY NOT NULL,
           type             text,
           description      text,
-          provider_type     text,
-          software         text,
+          provider_type    text,
           team             text,
           sources          text,
           status           text,
